# [CHIA-2052] optimize compressed CLVM serialization