--- a/src/serde/mod.rs
+++ b/src/serde/mod.rs
@@ -7,12 +7,15 @@ mod identity_hash;
 mod incremental;
 mod object_cache;
 mod parse_atom;
+mod path_builder;
 mod read_cache_lookup;
 mod ser;
 mod ser_br;
 mod serialized_length;
 mod tools;
+mod tree_cache;
 mod utils;
+mod visited_nodes;
 pub mod write_atom;
 
 #[cfg(test)]
@@ -26,6 +29,7 @@ pub use de_tree::{parse_triples, ParsedTriple};
 pub use identity_hash::RandomState;
 pub use incremental::{Serializer, UndoState};
 pub use object_cache::{serialized_length, treehash, ObjectCache};
+pub use path_builder::{ChildPos, PathBuilder};
 pub use read_cache_lookup::ReadCacheLookup;
 pub use ser::{node_to_bytes, node_to_bytes_limit};
 pub use ser_br::{node_to_bytes_backrefs, node_to_bytes_backrefs_limit};
@@ -33,3 +37,5 @@ pub use serialized_length::{serialized_length_atom, serialized_length_small_numb
 pub use tools::{
     serialized_length_from_bytes, serialized_length_from_bytes_trusted, tree_hash_from_stream,
 };
+pub use tree_cache::{TreeCache, TreeUndoState};
+pub use visited_nodes::VisitedNodes;